# collector/api/openapi.yaml
openapi: 3.0.0
info:
  title: LogMonitor Collector API
  description: API for LogMonitor frontend monitoring system
  version: 1.0.0
  contact:
    name: LogMonitor
    email: support@logmonitor.dev
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.logmonitor.dev
    description: Production server
...
paths:
...
  /api/admin/webhooks:
    post:
      tags: [Webhooks]
      summary: Create webhook
      description: Create a new webhook (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, url, events]
              properties:
                project_id:
                  type: integer
                  description: Project ID (0 for global webhooks)
                name:
                  type: string
                url:
                  type: string
                  format: uri
                secret:
                  type: string
                events:
                  type: array
                  items:
                    type: string
                    enum: [issue.created, issue.resolved, issue.reopened, alert.triggered]
      responses:
        '200':
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
...
  /api/admin/webhooks/{id}:
    delete:
      tags: [Webhooks]
      summary: Delete webhook
      description: Delete a webhook (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Webhook deleted
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Project API key for programmatic access
  schemas:
...
    Webhook:
      type: object
      properties:
        id:
          type: integer
        project_id:
          type: integer
        name:
          type: string
        url:
          type: string
          format: uri
        secret:
          type: string
        events:
          type: array
          items:
            type: string
            enum: [issue.created, issue.resolved, issue.reopened, alert.triggered]
        enabled:
          type: boolean
        last_triggered_at:
          type: integer
        failure_count:
          type: integer
        created_at:
          type: integer
        updated_at:
          type: integer